requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pymongo==4.8.0
python-dotenv==1.0.1
loguru==0.7.2